+++ restflow/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.2.0"
+__version__ = "1.3.0"
 __title__ = "Restflow"
 __author__ = "Khan Asfi Reza"
 __license__ = "BSD 3-Clause"